+++ Object/Character.cs
@@ -9,60 +9,61 @@ namespace spikewall.Object
 
         // The amount of rings this character
         // currently costs to level up.
-        public Int64? numRings { get; set; }
+        public long? numRings { get; set; }
 
         // UNUSED: The amount of Red Star Rings
         // this character costs to level up.
-        public Int64? numRedRings { get; set; }
+        public long? numRedRings { get; set; }
 
         // The amount of rings this
         // character costs to buy/limit smash.
-        public Int64? priceNumRings { get; set; }
+        public long? priceNumRings { get; set; }
 
         // The amount of Red Star Rings this
         // character costs to buy/limit smash.
-        public Int64? priceNumRedRings { get; set; }
+        public long? priceNumRedRings { get; set; }
 
         // Whether or not the character is unlocked.
-        public int? status { get; set; }
+        public sbyte? status { get; set; }
 
         // The level of the character.
-        public Int64? level { get; set; }
+        public sbyte? level { get; set; }
 
         // How many rings until
         // the next level up??
-        public Int64? exp { get; set; }
+        public long? exp { get; set; }
 
         // Amount of times the character
         // has been limit smashed.
-        public Int64? star { get; set; }
+        public sbyte? star { get; set; }
 
         // Maximum amount of times the
         // character can be limit smashed.
-        public Int64? starMax { get; set; }
+        public sbyte? starMax { get; set; }
 
         // How the character can be unlocked
         // (Purchasable with Red Rings, Winnable
         // on the Premium Roulette, etc).
-        public Int64? lockCondition { get; set; }
+        // TODO: Probably move this to an enum soon.
+        public sbyte? lockCondition { get; set; }
 
         // Not sure what this is right now.
         public Campaign[]? campaignList { get; set; }
 
         // The current levels for each ability.
-        public Int64[]? abilityLevel { get; set; }
+        public long[]? abilityLevel { get; set; }
 
         // Apparently this may be unused?
         // Otherwise, not sure what this is.
-        public Int64[]? abilityNumRings { get; set; }
+        public long[]? abilityNumRings { get; set; }
 
         // The current ability to be leveled up?
-        public Int64[]? abilityLevelup { get; set; }
+        public long[]? abilityLevelup { get; set; }
 
         // This is not always sent, so it needs to be
         // specifically handled during deserialization.
         // I'm also not sure what this is right now.
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-        public Int64[]? abilityLevelupExp { get; set; }
+        public long[]? abilityLevelupExp { get; set; }
     }
 }
